refacto(Safe): adapt safe tx handler to new structure
